requirements.txt
Django
Psycopg2
Pillow
Django-filter